requirements.txt
# ALPHARAGHU - Requirements (Windows Compatible)
alpaca-py>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
schedule>=1.2.0
aiohttp>=3.9.0
requests>=2.31.0
yfinance>=0.2.40
python-dotenv>=1.0.0
streamlit>=1.32.0
httpx>=0.27.0
